Added delete user usecase

+++ src/commonMain/kotlin/tz/co/asoft/auth/repo/AuthRepo.kt
@@ -1,40 +1,40 @@
 package tz.co.asoft.auth.repo
 
 import tz.co.asoft.auth.User
-import tz.co.asoft.auth.dao.AuthAbstractLocalDao
 import tz.co.asoft.auth.dao.IAuthDao
+import tz.co.asoft.auth.Email
+import tz.co.asoft.auth.Phone
+import tz.co.asoft.auth.dao.IAuthLocalDao
 import tz.co.asoft.io.file.File
 import tz.co.asoft.persist.dao.Dao
 import tz.co.asoft.persist.repo.Repo
 
-class AuthRepo(private val dao: Dao<User>, private val localDao: AuthAbstractLocalDao?) : Repo<User>(dao), IAuthRepo {
+class AuthRepo(private val dao: Dao<User>, private val localDao: Dao<User>?) : Repo<User>(dao), IAuthRepo {
 
-    override suspend fun emailSignIn(email: String, pwd: String) = if (dao is IAuthDao) {
-        dao.emailSignIn(email, pwd)?.also {
-            it.saveToLocalDb()
-        }
+    override suspend fun load(email: Email, pwd: String) = if (dao is IAuthDao) {
+        dao.load(email, pwd)?.also { it.saveToLocalDb() }
     } else null
 
-    override suspend fun phoneSignIn(phone: String, pwd: String) = if (dao is IAuthDao) {
-        dao.phoneSignIn(phone, pwd)?.also {
-            it.saveToLocalDb()
-        }
+    override suspend fun load(phone: Phone, pwd: String) = if (dao is IAuthDao) {
+        dao.load(phone, pwd)?.also { it.saveToLocalDb() }
     } else null
 
-    private suspend fun User.saveToLocalDb() = localDao?.save(this)
+    private suspend fun User.saveToLocalDb() = localDao?.create(this)
 
     override suspend fun uploadPhoto(user: User, photo: File) = if (dao is IAuthDao) {
         dao.uploadPhoto(user, photo)?.also {
-            localDao?.delete()
-            localDao?.save(it)
+            if (localDao is IAuthLocalDao?) localDao?.delete()
+            localDao?.create(it)
         }
     } else null
 
     override suspend fun saveToLocal(u: User) = u.saveToLocalDb()
-    override suspend fun loadLocalUser() = localDao?.load()
+    override suspend fun loadLocalUser() = if (localDao is IAuthLocalDao?) {
+        localDao?.load()
+    } else null
 
-    override suspend fun removeLocal() {
-        localDao?.delete()
+    override suspend fun deleteLocal() {
+        if (localDao is IAuthLocalDao?) localDao?.delete()
     }
 
     override suspend fun userWithEmailExists(emails: List<String>): Boolean = all()?.any { user ->
@@ -46,15 +46,20 @@ class AuthRepo(private val dao: Dao<User>, private val localDao: AuthAbstractLoc
     } ?: true
 
     override suspend fun edit(t: User): User? {
-        if (localDao?.load()?.uid == t.uid) {
-            localDao.save(t)
+        if (localDao is IAuthLocalDao?) {
+            if (localDao?.load()?.uid == t.uid) {
+                localDao?.delete()
+                localDao?.create(t)
+            }
         }
         return super.edit(t)
     }
 
     override suspend fun delete(t: User): User? {
-        if (localDao?.load()?.uid == t.uid) {
-            localDao.delete()
+        if (localDao is IAuthLocalDao?) {
+            if (localDao?.load()?.uid == t.uid) {
+                localDao?.delete()
+            }
         }
         return dao.delete(t)
     }
